move videoWindow to separate file
